Things are sort of a mess as we transition slowly to a modernized back-end architecture
Lots has been refactored into a namespace of smaller classes, the
primary of which delegates actions/hooks to others in the family.

This commit also introduces the Kirby Toolkit, to replace Underscore
and generally make the pains of PHP less severe— currently, the
templates and partials are sort of “half way” between the new and old
architecture, so very few pages work. Fortunately, the codebase has
been receptive to these updates, and there is a clear path forward.